requirements.txt
# Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Data & stats
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0
scikit-learn>=1.3.0

# ML & explainability
xgboost>=2.0.0
lightgbm>=4.2.0
shap>=0.44.0

# Frontend
streamlit>=1.29.0
plotly>=5.18.0
openpyxl>=3.1.0

# Optional: LLM for insight generation
# Anthropic Claude (preferred)
anthropic>=0.40.0
# OpenAI (fallback / alternative)
openai>=1.12.0
python-dotenv>=1.0.0
